Extract library body reference metadata resolver
