+++ docs/design/method-body-inspection.md
@@ -284,12 +284,16 @@ reader, generic scope, or raw IL. Call-site acquisition uses the same
 The assembly reader retains exactly: PE/body
 acquisition, the intentionally throwing `InstructionDecoder.Decode` +
 `BlockGraph.Build` decode, loop-region computation for the context, method
-identity and generic-scope creation, metadata ownership and token resolution,
-raw-IL ownership through the shared narrow method-analysis resolver, per-method
-orchestration and recoverable-failure diagnostics, resource/leak analysis, and
-result aggregation. Topic producers may each traverse the canonical decoded
-instructions for their own policy; this ownership split does not claim one
-instruction traversal overall.
+identity and generic-scope creation, primary-image metadata ownership and token
+resolution, raw-IL ownership through the shared narrow method-analysis resolver,
+per-method orchestration and recoverable-failure diagnostics, resource/leak
+analysis, and result aggregation. `LibraryBodyReferenceMetadataResolver`
+separately owns cross-assembly type-definition binding, referenced-image
+metadata lifetime, and its registration-keyed cache for that acquisition. It
+builds on `AssemblyReferenceBindingPolicy` and `TypeResolutionCatalog` rather
+than adding another binding engine. Topic producers may each traverse the
+canonical decoded instructions for their own policy; this ownership split does
+not claim one instruction traversal overall.
 `MethodInstructionFacts` owns the metadata-free local/argument-slot, operand,
 and single-branch-target grammar shared by safety and allocation interpretation,
 and `CompilerGeneratedNames` owns the unspeakable-name grammar shared by
